+++ src/slurm_plugin/common.py
@@ -12,7 +12,9 @@
 
 import functools
 import logging
+from concurrent.futures import Future
 from datetime import datetime
+from typing import Callable, Optional, Protocol, TypedDict
 
 from common.utils import check_command_output, time_is_up
 
@@ -22,6 +24,36 @@ logger = logging.getLogger(__name__)
 # YYYY-MM-DDTHH:MM:SS.ffffff+HH:MM[:SS[.ffffff]]
 TIMESTAMP_FORMAT = "%Y-%m-%d %H:%M:%S.%f%z"
 DEFAULT_COMMAND_TIMEOUT = 30
+
+ComputeInstanceDescriptor = TypedDict(
+    "ComputeInstanceDescriptor",
+    {
+        "Name": str,
+        "InstanceId": str,
+    },
+)
+
+
+class TaskController(Protocol):
+    class TaskShutdownError(RuntimeError):
+        """Exception raised if shutdown has been requested."""
+
+        pass
+
+    def queue_task(self, task: Callable[[], None]) -> Optional[Future]:
+        """Queue a task and returns a Future for the task or None if the task could not be queued."""
+
+    def is_shutdown(self) -> bool:
+        """Is shutdown has been requested."""
+
+    def raise_if_shutdown(self) -> None:
+        """Raise an error if a shutdown has been requested."""
+
+    def wait_unless_shutdown(self, seconds_to_wait: float) -> None:
+        """Wait for seconds_to_wait or will raise an error if a shutdown has been requested."""
+
+    def shutdown(self, wait: bool, cancel_futures: bool) -> None:
+        """Request that all tasks be shutdown."""
 
 
 def log_exception(