+++ tests/slurm_plugin/test_clustermgtd.py
@@ -14,13 +14,14 @@ import logging
 import os
 from datetime import datetime, timezone
 from types import SimpleNamespace
-from unittest.mock import call
+from unittest.mock import ANY, call
 
 import botocore
 import pytest
 import slurm_plugin
 from assertpy import assert_that
 from slurm_plugin.clustermgtd import ClusterManager, ClustermgtdConfig, ComputeFleetStatus, ComputeFleetStatusManager
+from slurm_plugin.console_logger import ConsoleLogger
 from slurm_plugin.fleet_manager import EC2Instance
 from slurm_plugin.slurm_resources import (
     EC2_HEALTH_STATUS_UNHEALTHY_STATES,
@@ -78,6 +79,13 @@ class TestClustermgtdConfig:
                     "health_check_timeout": 180,
                     "protected_failure_count": 10,
                     "insufficient_capacity_timeout": 600,
+                    # Compute console logging configs
+                    "compute_console_logging_enabled": True,
+                    "compute_console_logging_max_sample_size": 1,
+                    "compute_console_wait_time": 300,
+                    # Task executor configs
+                    "worker_pool_size": 5,
+                    "worker_pool_max_backlog": 100,
                 },
             ),
             (
@@ -111,6 +119,13 @@ class TestClustermgtdConfig:
                     "health_check_timeout": 10,
                     "protected_failure_count": 5,
                     "insufficient_capacity_timeout": 50.5,
+                    # Compute console logging configs
+                    "compute_console_logging_enabled": False,
+                    "compute_console_logging_max_sample_size": 50,
+                    "compute_console_wait_time": 10,
+                    # Task executor configs
+                    "worker_pool_size": 2,
+                    "worker_pool_max_backlog": 5,
                 },
             ),
             (
@@ -178,12 +193,21 @@ class TestClustermgtdConfig:
         assert_that(ClustermgtdConfig(config)).is_not_equal_to(ClustermgtdConfig(config_modified))
 
 
+@pytest.mark.usefixtures("initialize_console_logger_mock")
 def test_set_config(initialize_instance_manager_mock):
     initial_config = SimpleNamespace(
-        some_key_1="some_value_1", some_key_2="some_value_2", insufficient_capacity_timeout=20
+        some_key_1="some_value_1",
+        some_key_2="some_value_2",
+        insufficient_capacity_timeout=20,
+        worker_pool_size=5,
+        worker_pool_max_backlog=10,
     )
     updated_config = SimpleNamespace(
-        some_key_1="some_value_1", some_key_2="some_value_2_changed", insufficient_capacity_timeout=10
+        some_key_1="some_value_1",
+        some_key_2="some_value_2_changed",
+        insufficient_capacity_timeout=10,
+        worker_pool_size=10,
+        worker_pool_max_backlog=5,
     )
 
     cluster_manager = ClusterManager(initial_config)
@@ -194,6 +218,40 @@ def test_set_config(initialize_instance_manager_mock):
     assert_that(cluster_manager._config).is_equal_to(updated_config)
 
     assert_that(initialize_instance_manager_mock.call_count).is_equal_to(2)
+
+
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
+def test_exception_from_report_console_output_from_nodes(mocker):
+    config = SimpleNamespace(
+        some_key_1="some_value_1",
+        some_key_2="some_value_2",
+        insufficient_capacity_timeout=20,
+        worker_pool_size=5,
+        worker_pool_max_backlog=0,
+        compute_console_logging_max_sample_size=2,
+        compute_console_wait_time=10,
+    )
+    unhealthy_nodes = [
+        StaticNode("queue1-st-c5xlarge-3", "ip-3", "hostname", "some_state", "queue1"),
+    ]
+    reset_nodes_mock = mocker.patch("slurm_plugin.clustermgtd.reset_nodes", autospec=True)
+    cluster_manager = ClusterManager(config)
+
+    mock_console_logger = mocker.patch.object(cluster_manager, "_console_logger", spec=ConsoleLogger)
+    report_mock = mock_console_logger.report_console_output_from_nodes
+    report_mock.side_effect = Exception
+
+    cluster_manager._handle_unhealthy_static_nodes(unhealthy_nodes)
+
+    # Make sure report_console_output_from_nodes was called.
+    report_mock.assert_called_once()
+
+    # Make sure the code after _report_console_from_output_nodes is called.
+    reset_nodes_mock.assert_called_once_with(
+        ANY, state="down", reason="Static node maintenance: unhealthy node is being replaced"
+    )
 
 
 @pytest.mark.parametrize(
@@ -263,8 +321,10 @@ def test_set_config(initialize_instance_manager_mock):
     ],
     ids=["normal", "delete_exception", "reset_exception"],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
-def test_clean_up_inactive_parititon(
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
+def test_clean_up_inactive_partition(
     inactive_instances,
     slurm_inactive_nodes,
     expected_reset,
@@ -299,6 +359,7 @@ def test_clean_up_inactive_parititon(
         mock_reset_node.assert_not_called()
 
 
+@pytest.mark.usefixtures("initialize_executor_mock", "initialize_console_logger_mock")
 def test_get_ec2_instances(mocker):
     # Test setup
     mock_sync_config = SimpleNamespace(
@@ -470,6 +531,7 @@ def test_get_ec2_instances(mocker):
     ],
     ids=["basic", "disable_ec2", "disable_all", "disable_scheduled", "no_unhealthy_instance"],
 )
+@pytest.mark.usefixtures("initialize_executor_mock", "initialize_console_logger_mock")
 def test_perform_health_check_actions(
     mock_instance_health_states,
     disable_ec2_health_check,
@@ -551,7 +613,9 @@ def test_perform_health_check_actions(
     ],
     ids=["scheduled_event", "ec2_health", "all_healthy"],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_handle_health_check(
     health_check_type,
     mock_fail_ec2_side_effect,
@@ -623,7 +687,9 @@ def test_handle_health_check(
     ],
     ids=["mixed"],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_update_static_nodes_in_replacement(current_replacing_nodes, slurm_nodes, expected_replacing_nodes, mocker):
     mock_sync_config = SimpleNamespace(insufficient_capacity_timeout=600)
     cluster_manager = ClusterManager(mock_sync_config)
@@ -716,7 +782,9 @@ def test_update_static_nodes_in_replacement(current_replacing_nodes, slurm_nodes
     ],
     ids=["basic", "disable smart instance capacity"],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_handle_unhealthy_dynamic_nodes(
     unhealthy_dynamic_nodes,
     instances,
@@ -772,7 +840,9 @@ def test_handle_unhealthy_dynamic_nodes(
     ],
     ids=["basic"],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_handle_powering_down_nodes(
     slurm_nodes, instances, instances_to_terminate, expected_powering_down_nodes, mocker
 ):
@@ -796,7 +866,9 @@ def test_handle_powering_down_nodes(
         "expected_replacing_nodes",
         "delete_instance_list",
         "add_node_list",
+        "output_enabled",
         "set_nodes_down_exception",
+        "sample_size",
     ),
     [
         (
@@ -824,7 +896,9 @@ def test_handle_powering_down_nodes(
             },
             ["id-1", "id-2"],
             ["queue1-st-c5xlarge-1", "queue1-st-c5xlarge-2", "queue1-st-c5xlarge-3"],
+            True,
             None,
+            100,
         ),
         (
             {"current-queue1-st-c5xlarge-6"},
@@ -847,7 +921,9 @@ def test_handle_powering_down_nodes(
             },
             [],
             ["queue1-st-c5xlarge-1", "queue1-st-c5xlarge-2", "queue1-st-c5xlarge-3"],
+            False,
             None,
+            1,
         ),
         (
             {"current-queue1-st-c5xlarge-6"},
@@ -864,7 +940,9 @@ def test_handle_powering_down_nodes(
             {"current-queue1-st-c5xlarge-6", "queue1-st-c5xlarge-1", "queue1-st-c5xlarge-2"},
             [],
             ["queue1-st-c5xlarge-1", "queue1-st-c5xlarge-2", "queue1-st-c5xlarge-3"],
+            False,
             None,
+            1,
         ),
         (
             {"current-queue1-st-c5xlarge-6"},
@@ -881,7 +959,9 @@ def test_handle_powering_down_nodes(
             {"current-queue1-st-c5xlarge-6", "queue1-st-c5xlarge-1", "queue1-st-c5xlarge-2"},
             [],
             ["queue1-st-c5xlarge-1", "queue1-st-c5xlarge-2", "queue1-st-c5xlarge-3"],
+            False,
             Exception,
+            1,
         ),
     ],
     ids=["basic", "no_associated_instances", "partial_launch", "set_nodes_down_exception"],
@@ -897,7 +977,9 @@ def test_handle_unhealthy_static_nodes(
     mocker,
     caplog,
     request,
+    output_enabled,
     set_nodes_down_exception,
+    sample_size,
 ):
     # Test setup
     mock_sync_config = SimpleNamespace(
@@ -918,6 +1000,11 @@ def test_handle_unhealthy_static_nodes(
         insufficient_capacity_timeout=600,
         run_instances_overrides={},
         create_fleet_overrides={},
+        compute_console_logging_enabled=output_enabled,
+        compute_console_logging_max_sample_size=sample_size,
+        compute_console_wait_time=1,
+        worker_pool_size=5,
+        worker_pool_max_backlog=10,
     )
     for node, instance in zip(unhealthy_static_nodes, instances):
         node.instance = instance
@@ -937,6 +1024,11 @@ def test_handle_unhealthy_static_nodes(
             }
         }
     )
+
+    report_mock = mocker.patch.object(
+        cluster_manager._console_logger, "report_console_output_from_nodes", autospec=True
+    )
+
     mocker.patch("slurm_plugin.instance_manager.update_nodes")
     cluster_manager._instance_manager._store_assigned_hostnames = mocker.MagicMock()
     cluster_manager._instance_manager._update_dns_hostnames = mocker.MagicMock()
@@ -956,6 +1048,7 @@ def test_handle_unhealthy_static_nodes(
 
     # Run test
     cluster_manager._handle_unhealthy_static_nodes(unhealthy_static_nodes)
+
     if set_nodes_down_exception is Exception:
         assert_that(caplog.text).contains("Encountered exception when setting unhealthy static nodes into down state:")
         return
@@ -973,6 +1066,7 @@ def test_handle_unhealthy_static_nodes(
     if "partial_launch" not in request.node.name:
         assert_that(caplog.text).is_empty()
     assert_that(cluster_manager._static_nodes_in_replacement).is_equal_to(expected_replacing_nodes)
+    report_mock.assert_called_once()
 
 
 @pytest.mark.parametrize(
@@ -1086,7 +1180,9 @@ def test_handle_unhealthy_static_nodes(
     ],
     ids=["basic", "disable smart instance capacity fail over mechanism"],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_maintain_nodes(
     active_nodes,
     instances,
@@ -1182,6 +1278,9 @@ def test_maintain_nodes(
         ),
     ],
     ids=["all_good", "orphaned", "orphaned_timeout"],
+)
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
 )
 def test_terminate_orphaned_instances(
     cluster_instances, slurm_nodes, current_time, expected_instance_to_terminate, mocker
@@ -1335,6 +1434,7 @@ def test_terminate_orphaned_instances(
         "stopped_disabled",
     ],
 )
+@pytest.mark.usefixtures("initialize_executor_mock", "initialize_console_logger_mock")
 def test_manage_cluster(
     disable_cluster_management,
     disable_health_check,
@@ -1821,6 +1921,7 @@ def test_manage_cluster(
     ],
     ids=["basic", "failures", "critical_failure_1", "critical_failure_2"],
 )
+@pytest.mark.usefixtures("initialize_executor_mock", "initialize_console_logger_mock")
 def test_manage_cluster_boto3(
     boto3_stubber,
     config_file,
@@ -1945,6 +2046,7 @@ class TestComputeFleetStatusManager:
         ),
     ],
 )
+@pytest.mark.usefixtures("initialize_executor_mock", "initialize_console_logger_mock")
 def test_handle_successfully_launched_nodes(
     healthy_nodes,
     expected_partitions_protected_failure_count_map,
@@ -2023,7 +2125,9 @@ def test_handle_successfully_launched_nodes(
         ),
     ],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_increase_partitions_protected_failure_count(nodes, initial_map, expected_map, mocker):
     # Mock associated function
     mock_sync_config = SimpleNamespace(insufficient_capacity_timeout=600)
@@ -2042,7 +2146,9 @@ def test_increase_partitions_protected_failure_count(nodes, initial_map, expecte
         ("queue2", {"queue1": 2}),
     ],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_reset_partition_failure_count(mocker, partition, expected_map):
     # Mock associated function
     mock_sync_config = SimpleNamespace(insufficient_capacity_timeout=600)
@@ -2099,7 +2205,9 @@ def test_reset_partition_failure_count(mocker, partition, expected_map):
         ),
     ],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_handle_protected_mode_process(
     partitions,
     slurm_nodes_list,
@@ -2149,7 +2257,9 @@ def test_handle_protected_mode_process(
         ),
     ],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_enter_protected_mode(
     partitions_to_disable,
     mocker,
@@ -2175,6 +2285,20 @@ def test_enter_protected_mode(
 def initialize_instance_manager_mock(mocker):
     return mocker.patch.object(
         ClusterManager, "_initialize_instance_manager", spec=ClusterManager._initialize_instance_manager
+    )
+
+
+@pytest.fixture()
+def initialize_executor_mock(mocker):
+    return mocker.patch.object(ClusterManager, "_initialize_executor", spec=ClusterManager._initialize_executor)
+
+
+@pytest.fixture()
+def initialize_console_logger_mock(mocker):
+    return mocker.patch.object(
+        ClusterManager,
+        "_initialize_console_logger",
+        spec=ClusterManager._initialize_console_logger,
     )
 
 
@@ -2212,7 +2336,9 @@ def initialize_instance_manager_mock(mocker):
     ],
     ids=["not_in_replacement", "no-backing-instance", "in_replacement", "timeout"],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_is_node_being_replaced(current_replacing_nodes, node, instance, current_time, expected_result):
     mock_sync_config = SimpleNamespace(node_replacement_timeout=30, insufficient_capacity_timeout=3)
     cluster_manager = ClusterManager(mock_sync_config)
@@ -2263,7 +2389,9 @@ def test_is_node_being_replaced(current_replacing_nodes, node, instance, current
         ),
     ],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_is_node_replacement_timeout(node, current_node_in_replacement, is_replacement_timeout, instance):
     node.instance = instance
     mock_sync_config = SimpleNamespace(node_replacement_timeout=30, insufficient_capacity_timeout=-2.2)
@@ -2314,7 +2442,9 @@ def test_is_node_replacement_timeout(node, current_node_in_replacement, is_repla
         ),
     ],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_handle_failed_health_check_nodes_in_replacement(
     active_nodes,
     is_static_nodes_in_replacement,
@@ -2375,7 +2505,9 @@ def test_handle_failed_health_check_nodes_in_replacement(
         ),
     ],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_handle_bootstrap_failure_nodes(
     active_nodes,
     instances,
@@ -2437,7 +2569,9 @@ def test_handle_bootstrap_failure_nodes(
     ],
     ids=["basic"],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_find_bootstrap_failure_nodes(active_nodes, instances):
     # Mock functions
     mock_sync_config = SimpleNamespace(
@@ -2595,7 +2729,9 @@ def test_find_bootstrap_failure_nodes(active_nodes, instances):
         ),
     ],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_handle_ice_nodes(
     queue_compute_resource_nodes_map,
     ice_compute_resources_and_nodes_map,
@@ -2758,7 +2894,9 @@ def test_handle_ice_nodes(
         ),
     ],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_update_insufficient_capacity_compute_resources(
     ice_compute_resources_and_nodes_map,
     initial_insufficient_capacity_compute_resources,
@@ -2927,7 +3065,9 @@ def test_update_insufficient_capacity_compute_resources(
         ),
     ],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_reset_timeout_expired_compute_resources(
     ice_compute_resources_and_nodes_map,
     initial_insufficient_capacity_compute_resources,
@@ -3030,7 +3170,9 @@ def test_reset_timeout_expired_compute_resources(
         ),
     ],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_set_ice_compute_resources_to_down(
     queue_compute_resource_nodes_map,
     insufficient_capacity_compute_resources,
@@ -3267,7 +3409,9 @@ def test_set_ice_compute_resources_to_down(
         ),
     ],
 )
-@pytest.mark.usefixtures("initialize_instance_manager_mock")
+@pytest.mark.usefixtures(
+    "initialize_instance_manager_mock", "initialize_executor_mock", "initialize_console_logger_mock"
+)
 def test_find_unhealthy_slurm_nodes(
     active_nodes,
     expected_unhealthy_dynamic_nodes,